fix(admin): resolve SQL errors and improve artist search UX
- Fix admin locations API to use artist_locations table (artists.city/state dropped)
- Fix get_artists_with_image_counts return type (follower_count integer not bigint)
- Add get_artist_stats function to admin_functions.sql
- Strip @ prefix from admin artist search input
- Link @handle in artist table to admin artist page
- Use centralized generateSlugFromInstagram() in discovery scripts to prevent
  invalid slugs with leading/trailing hyphens

+++ supabase/functions/admin/admin_functions.sql
@@ -138,7 +138,7 @@ RETURNS TABLE (
   is_featured boolean,
   is_pro boolean,
   verification_status text,
-  follower_count bigint,
+  follower_count integer,
   slug text,
   deleted_at timestamp with time zone,
   image_count bigint,
